requirements.txt
# Core dependencies for the AI microservices
langchain==0.2.0
langchain-community==0.2.0
langchain-core==0.2.0
# flowise==1.0.0  # Commented out - install separately with npm

# For document processing
pypdf==4.0.0
python-docx==0.8.11

# Web framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# For API clients
requests==2.31.0

# Environment management
python-dotenv==1.0.0

# For vector stores (if needed)
chromadb==0.4.22
sentence-transformers==2.2.2

# Utilities
pydantic==2.5.0

# Authentication and security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
SQLAlchemy==2.0.23
alembic==1.13.1

# CORS support
fastapi-cors==0.0.6

# Additional dependencies for better functionality
huggingface-hub==0.19.4
transformers==4.36.0